+++ CHANGELOG.md
@@ -25,6 +25,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Changed serialization of security-events.
+  [#540](https://github.com/astarte-platform/astarte-device-sdk-rust/pull/540)
+
 ## [v0.11.0] - 2025-12-18
 
 ### Added